Move DEFAULT_PERSONAL_VELP_GROUP_NAME from defaultconfig.py to timApp/velp/velp.py to prevent accidental modification
